Changed Location model

+++ src/main/java/gut/follower/org/Models/Location.java
@@ -2,27 +2,27 @@ package gut.follower.org.Models;
 
 public class Location {
 
-    private Long latitude;
-    private Long longitude;
+    private double latitude;
+    private double longitude;
 
-    public Location(Long latitude, Long longitude) {
+    public Location(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public Long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
